Refactor document/Access to common component, use in Project header
