+++ src/lib/components/documents/DocumentListItem.svelte
@@ -13,8 +13,7 @@ If we're in an embed, we want to open links to documents in new tabs and hide th
   import { _ } from "svelte-i18n";
   import { Alert24, Hourglass24, File24 } from "svelte-octicons";
 
-  import DocAccess, { getLevel } from "../documents/Access.svelte";
-  import Flex from "../common/Flex.svelte";
+  import DocAccess, { getLevel } from "../common/Access.svelte";
   import KV from "../common/KV.svelte";
   import NoteTab from "../viewer/NoteTab.svelte";
